Add combineAppRoutes util
This update adds a new combineAppRoutes utility function. This enables Marionette
apps to render a React Router powered component/app. It works by combining React
routes and Marionette Routes to a flat object for Marionette's appRouter.

The trick is to have the React routes end with `*notFound`, which kind of functions
like a wild card match.

It's an unfortunate thing, but it's the only thing that seems to work.